Close the P0 gate: build the tree, run the UI live, fix what that found
scons was available all along -- the submodules were simply not checked out,
which is why the earlier note said there was no build. With msgq_repo,
rednose_repo and opendbc_repo initialised and capnproto/eigen/OpenCL headers
installed, cereal, msgq, visionipc and common build, and the UI runs against
real msgq, real capnp and real Params. Twelve frames rendered, no errors,
every value correct.

Running it found five defects that no amount of testing against a stub would
have caught.

SubMaster.valid is a dict, not a method. state.py called sm.valid() at 24
sites, so _read() raised TypeError on the very first tick against a real
SubMaster -- the UI would never have drawn a single frame on the device. The
test double had valid as a method, which is precisely what hid it; it is a
dict now, and its __getitem__ returns a default message for an unseen service
the way the real one does rather than raising.

deviceState.networkType, networkStrength and thermalStatus are capnp enums.
int() on a _DynamicEnum raises. They are read by name now, with an explicit
name -> level table for strength so an enum member added later reads as no
signal rather than as full bars.

started was gated on deviceState being valid. ui.cc does not gate it, and it
should not: whether the car is on is not something to suppress because some
other field in deviceState is degraded.

ApiCache_DriveStats and OffroadAlerts do not exist. Params raises
UnknownKeyName for an undeclared key, so the home screen crashed on open.
Neither was ever going to work: this fork keeps its own EOPTrip* counters
rather than an API cache, and offroad alerts are one Offroad_* JSON param per
alert with the catalogue shipped in selfdrived/alerts_offroad.json -- which is
how offroad_alerts.cc reads them. Both rewritten against the real keys, and
every one of the 37 param keys the UI touches is now checked against
params_keys.h.

Two build-blocking bugs in cereal, both of which would stop any clang build:

  services.h declared `int frequency` while the generator writes Python
  floats into the initialiser. clang rejects the narrowing outright. Casting
  would have been the wrong fix -- four services are slower than 1 Hz
  (thumbnail 0.0167, carParams 0.02, clocks 0.1, procLog 0.5) and would
  truncate to 0.

  socketmaster's SubMessage had the same `int freq`. Its liveness check is
  `freq <= 1e-5 || elapsed < 10.0 / freq` -- a float comparison against an
  int, and with truncation those same four services take the "no rate to
  check" branch and are reported alive forever whether or not they ever
  arrive.

The side and rear camera overlays are no longer placeholders. They are real
VisionIPC views on uvcd's SIDE_LEFT/SIDE_RIGHT/REAR streams -- uvcd, not
camerad, because they are ordinary UVC devices rather than the ISP pipeline.
They take the CPU path deliberately: they appear and disappear mid-drive, and
EGLFS permits one fullscreen GL surface, so a second QOpenGLWidget showing up
when the driver signals is the case that terminates the application. Only the
overlay that is actually on screen is polled, and BGR is byte-swapped rather
than reinterpreted -- getting that backwards is not a crash, it is a picture
with red and blue exchanged.

115 tests pass on PyQt5 and PySide6; ./test.sh --no-pytest is green.

+++ cereal/messaging/socketmaster.cc
@@ -33,7 +33,14 @@ MessageContext message_context;
 struct SubMaster::SubMessage {
   std::string name;
   SubSocket *socket = nullptr;
-  int freq = 0;
+  // float, not int. Four services are slower than 1 Hz -- thumbnail at
+  // 0.0167, carParams at 0.02, clocks at 0.1, procLog at 0.5 -- and an int
+  // truncates every one of them to 0. The liveness check below then takes
+  // the `freq <= 1e-5` branch, which means "no rate to check against", so
+  // those four would be reported alive forever whether or not they ever
+  // arrived. That 1e-5 comparison is itself against a float, which is what
+  // this field was always meant to be.
+  float freq = 0;
   bool updated = false, alive = false, valid = false, ignore_alive;
   uint64_t rcv_time = 0, rcv_frame = 0;
   void *allocated_msg_reader = nullptr;
